+++ package.json
@@ -80,7 +80,7 @@
     "ncp": "^2.0.0",
     "next-compose-plugins": "^2.2.0",
     "next-images": "^1.1.2",
-    "npm": "^6.10.1",
+    "npm": "^6.13.4",
     "prettier": "^1.18.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
