+++ pages/index.tsx
@@ -11,6 +11,7 @@ import { ISettings } from 'common/interfaces';
 import { settingsReducer } from 'store';
 import ToolBar from 'components/ToolBar';
 // import { ReactMultiEmail } from '@axui/react-multi-email/ReactMultiEmail';
+import ReactMultiEmail from '../packages/react-multi-email/ReactMultiEmail';
 
 const initialSettings: ISettings = {
   width: 300,
@@ -21,6 +22,7 @@ const Home: React.FC = props => {
     settingsReducer,
     initialSettings,
   );
+  const [emails, setEmails] = useState();
 
   useEffect(() => {
     //
@@ -31,7 +33,28 @@ const Home: React.FC = props => {
       Hello World!
       <br />
       width: {settings.width}, height: {settings.height}
-      <ToolBar {...settings} dispatchSettings={dispatchSettings} />
+      <ReactMultiEmail
+        placeholder="Input your email"
+        emails={emails}
+        onChange={(_emails: string[]) => {
+          console.log(_emails);
+          setEmails({ emails: _emails });
+        }}
+        getLabel={(
+          email: string,
+          index: number,
+          removeEmail: (index: number) => void,
+        ) => {
+          return (
+            <div data-tag key={index}>
+              <div data-tag-item>{email}</div>
+              <span data-tag-handle onClick={() => removeEmail(index)}>
+                ×
+              </span>
+            </div>
+          );
+        }}
+      />
     </div>
   );
 };
